src/EMRNext.Core/Reporting/Services/PredictiveAnalyticsService.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;
using Microsoft.Extensions.Logging;
using Microsoft.ML;
using Microsoft.ML.Data;
using EMRNext.Core.Domain.Entities;
using EMRNext.Core.Repositories;

namespace EMRNext.Core.Reporting.Services
{
    /// <summary>
    /// Predictive analytics service for medical insights
    /// </summary>
    public class PredictiveAnalyticsService
    {
        private readonly ILogger<PredictiveAnalyticsService> _logger;
        private readonly IGenericRepository<Patient> _patientRepository;
        private readonly MLContext _mlContext;

        public PredictiveAnalyticsService(
            ILogger<PredictiveAnalyticsService> logger,
            IGenericRepository<Patient> patientRepository)
        {
            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
            _patientRepository = patientRepository ?? throw new ArgumentNullException(nameof(patientRepository));
            _mlContext = new MLContext(seed: 0);
        }

        /// <summary>
        /// Predict disease risk for a patient
        /// </summary>
        public async Task<DiseaseRiskPrediction> PredictDiseaseRiskAsync(Guid patientId)
        {
            try 
            {
                var patient = await _patientRepository.GetByIdAsync(patientId);
                
                if (patient == null)
                    throw new ArgumentException("Patient not found");

                var patientData = PreparePatientData(patient);
                var predictionEngine = LoadDiseaseRiskModel();

                var prediction = predictionEngine.Predict(patientData);

                return new DiseaseRiskPrediction
                {
                    PatientId = patientId,
                    DiabetesRisk = prediction.DiabetesRisk,
                    HeartDiseaseRisk = prediction.HeartDiseaseRisk,
                    HypertensionRisk = prediction.HypertensionRisk,
                    PredictionConfidence = prediction.Confidence
                };
            }
            catch (Exception ex)
            {
                _logger.LogError(ex, $"Error predicting disease risk for patient {patientId}");
                throw;
            }
        }

        /// <summary>
        /// Train disease risk prediction model
        /// </summary>
        public async Task<ModelTrainingResult> TrainDiseaseRiskModelAsync()
        {
            try 
            {
                var patients = await _patientRepository.GetAllAsync();
                var trainingData = patients
                    .Select(PreparePatientData)
                    .ToList();

                var dataView = _mlContext.Data.LoadFromEnumerable(trainingData);

                var pipeline = _mlContext.Transforms.Concatenate(
                        "Features", 
                        "Age", "Weight", "Height", "BloodPressure", "CholesterolLevel"
                    )
                    .Append(_mlContext.BinaryClassification.Trainers.SdcaLogisticRegression());

                var model = pipeline.Fit(dataView);

                // Save model
                _mlContext.Model.Save(model, dataView.Schema, "disease_risk_model.zip");

                return new ModelTrainingResult
                {
                    ModelPath = "disease_risk_model.zip",
                    TrainingDate = DateTime.UtcNow,
                    NumberOfPatients = trainingData.Count
                };
            }
            catch (Exception ex)
            {
                _logger.LogError(ex, "Error training disease risk prediction model");
                throw;
            }
        }

        /// <summary>
        /// Predict patient treatment adherence
        /// </summary>
        public async Task<TreatmentAdherencePrediction> PredictTreatmentAdherenceAsync(Guid patientId)
        {
            try 
            {
                var patient = await _patientRepository.GetByIdAsync(patientId);
                
                if (patient == null)
                    throw new ArgumentException("Patient not found");

                var patientData = PreparePatientAdherenceData(patient);
                var predictionEngine = LoadTreatmentAdherenceModel();

                var prediction = predictionEngine.Predict(patientData);

                return new TreatmentAdherencePrediction
                {
                    PatientId = patientId,
                    AdherenceProbability = prediction.AdherenceProbability,
                    PredictionConfidence = prediction.Confidence
                };
            }
            catch (Exception ex)
            {
                _logger.LogError(ex, $"Error predicting treatment adherence for patient {patientId}");
                throw;
            }
        }

        /// <summary>
        /// Prepare patient data for disease risk prediction
        /// </summary>
        private PatientRiskData PreparePatientData(Patient patient)
        {
            return new PatientRiskData
            {
                Age = patient.Age,
                Weight = patient.Weight,
                Height = patient.Height,
                BloodPressure = patient.BloodPressure,
                CholesterolLevel = patient.CholesterolLevel,
                // Add more relevant health indicators
            };
        }

        /// <summary>
        /// Prepare patient data for treatment adherence prediction
        /// </summary>
        private PatientAdherenceData PreparePatientAdherenceData(Patient patient)
        {
            return new PatientAdherenceData
            {
                Age = patient.Age,
                PreviousAdherenceRate = CalculatePreviousAdherenceRate(patient),
                NumberOfMedications = patient.Prescriptions?.Count ?? 0,
                ChronicConditionsCount = CountChronicConditions(patient)
            };
        }

        /// <summary>
        /// Load pre-trained disease risk prediction model
        /// </summary>
        private PredictionEngine<PatientRiskData, DiseaseRiskPredictionResult> LoadDiseaseRiskModel()
        {
            var model = _mlContext.Model.Load("disease_risk_model.zip", out var modelSchema);
            return _mlContext.Model.CreatePredictionEngine<PatientRiskData, DiseaseRiskPredictionResult>(model);
        }

        /// <summary>
        /// Load pre-trained treatment adherence model
        /// </summary>
        private PredictionEngine<PatientAdherenceData, TreatmentAdherencePredictionResult> LoadTreatmentAdherenceModel()
        {
            var model = _mlContext.Model.Load("treatment_adherence_model.zip", out var modelSchema);
            return _mlContext.Model.CreatePredictionEngine<PatientAdherenceData, TreatmentAdherencePredictionResult>(model);
        }

        /// <summary>
        /// Calculate previous treatment adherence rate
        /// </summary>
        private double CalculatePreviousAdherenceRate(Patient patient)
        {
            // Implement logic to calculate adherence rate from historical data
            return 0.75; // Placeholder
        }

        /// <summary>
        /// Count number of chronic conditions
        /// </summary>
        private int CountChronicConditions(Patient patient)
        {
            // Implement logic to count chronic conditions
            return 0; // Placeholder
        }
    }

    /// <summary>
    /// Patient risk data for ML prediction
    /// </summary>
    public class PatientRiskData
    {
        [LoadColumn(0)]
        public float Age { get; set; }

        [LoadColumn(1)]
        public float Weight { get; set; }

        [LoadColumn(2)]
        public float Height { get; set; }

        [LoadColumn(3)]
        public float BloodPressure { get; set; }

        [LoadColumn(4)]
        public float CholesterolLevel { get; set; }
    }

    /// <summary>
    /// Patient adherence data for ML prediction
    /// </summary>
    public class PatientAdherenceData
    {
        [LoadColumn(0)]
        public float Age { get; set; }

        [LoadColumn(1)]
        public float PreviousAdherenceRate { get; set; }

        [LoadColumn(2)]
        public float NumberOfMedications { get; set; }

        [LoadColumn(3)]
        public float ChronicConditionsCount { get; set; }
    }

    /// <summary>
    /// Disease risk prediction result
    /// </summary>
    public class DiseaseRiskPredictionResult
    {
        [ColumnName("PredictedLabel")]
        public bool HasHighRisk { get; set; }

        [ColumnName("Probability")]
        public float Confidence { get; set; }

        public float DiabetesRisk { get; set; }
        public float HeartDiseaseRisk { get; set; }
        public float HypertensionRisk { get; set; }
    }

    /// <summary>
    /// Treatment adherence prediction result
    /// </summary>
    public class TreatmentAdherencePredictionResult
    {
        [ColumnName("PredictedLabel")]
        public bool WillAdhere { get; set; }

        [ColumnName("Probability")]
        public float Confidence { get; set; }

        public float AdherenceProbability { get; set; }
    }

    /// <summary>
    /// Disease risk prediction for a patient
    /// </summary>
    public class DiseaseRiskPrediction
    {
        public Guid PatientId { get; set; }
        public float DiabetesRisk { get; set; }
        public float HeartDiseaseRisk { get; set; }
        public float HypertensionRisk { get; set; }
        public float PredictionConfidence { get; set; }
    }

    /// <summary>
    /// Treatment adherence prediction for a patient
    /// </summary>
    public class TreatmentAdherencePrediction
    {
        public Guid PatientId { get; set; }
        public float AdherenceProbability { get; set; }
        public float PredictionConfidence { get; set; }
    }

    /// <summary>
    /// Result of model training
    /// </summary>
    public class ModelTrainingResult
    {
        public string ModelPath { get; set; }
        public DateTime TrainingDate { get; set; }
        public int NumberOfPatients { get; set; }
    }
}
